requirements.txt
Flask
Flask-Cors
supabase
python-dotenv
PyJWT
bcrypt
Flask-Mail
pinecone
together
langchain==0.1.16
langchain-core==0.1.52
langchain-community==0.0.32
langchain-pinecone==0.1.3
langchain-together==0.1.2
